Changes to SLAM, Changed Teleop Speed, Changed Laserscan Rate
Still working on tuning SLAM. The lidar rate was reduced from 50 hz to 10 hz which is the same as the RPLidar A2. This reduced rate seems to increase the likelihood of losing track of position and decreases SLAM quality. The turn rate was increased slightly, the forward speed was reduced to a more reasonable level.