Improved swerve simulation.
Fixed code and settings that made the swerve not want to do anything,
Still needs improvement. Having trouble with field-oriented. Still no Odometry from the wheels, however this has been added to the real robot code and I might steal that.
Also it's 2 AM on a Thursday. Thank God i am working from home tomorrow. The 2020 Kickoff is in 32 hours. It's gonna be a tough season.